Add CsWin32 for Native Methods and Disable Marshaling
Move All Native Methods to a Single NativeMehtod.cs class
